+++ batch-api/batch_request_builder2.go
@@ -11,26 +11,31 @@ const (
 	batchURLTemplate = "{+baseurl}/v1/batch" // Must be versioned
 )
 
-// BatchRequestBuilder constructs batch requests for the specified base URL.
-type BatchRequestBuilder2 struct {
+// BatchRequestBuilder2 defines the operations available for the batch api as a whole.
+type BatchRequestBuilder2 interface {
+	Post(ctx context.Context, request BatchRequest) (BatchResponse, error)
+}
+
+// batchRequestBuilder2 constructs batch requests for the specified base URL.
+type batchRequestBuilder2 struct {
 	intCore.Sendable
 }
 
 // NewBatchRequestBuilder creates a new BatchRequestBuilder.
 // It takes a client (core.Client) and pathParameters (map[string]string).
-func NewBatchRequestBuilder2(client intCore.ClientSendable, pathParameters map[string]string) *BatchRequestBuilder2 {
+func NewBatchRequestBuilder2(client intCore.ClientSendable, pathParameters map[string]string) (BatchRequestBuilder2, error) {
 	requestBuilder := intCore.NewRequestBuilder2(
 		client,
 		batchURLTemplate,
 		pathParameters,
 	)
-	return &BatchRequestBuilder2{
+	return &batchRequestBuilder2{
 		requestBuilder,
-	}
+	}, nil
 }
 
 // Post sends a batch request and returns the BatchResponse.
-func (rB *BatchRequestBuilder2) Post(ctx context.Context, request BatchRequest) (BatchResponse, error) {
+func (rB *batchRequestBuilder2) Post(ctx context.Context, request BatchRequest) (BatchResponse, error) {
 	config := &core.RequestConfiguration{
 		Data:     request,
 		Response: &batchResponse{},